Use Attachments API implementation provided by interactive-api-host package
[#179181324]
