+++ package.json
@@ -145,6 +145,7 @@
   "dependencies": {
     "@aws-sdk/client-s3": "^3.24.0",
     "@aws-sdk/s3-request-presigner": "^3.24.0",
+    "@concord-consortium/interactive-api-host": "0.1.0-pre.3",
     "@concord-consortium/lara-interactive-api": "^1.1.2",
     "@concord-consortium/text-decorator": "^1.0.2",
     "@concord-consortium/token-service": "^2.1.0",
