+++ src/utilities/get-attachments-manager-options.ts
@@ -1,8 +1,10 @@
+import { IAttachmentsManagerInitOptions } from "@concord-consortium/interactive-api-host";
 import { EnvironmentName } from "@concord-consortium/token-service";
 import { firebaseAppName, getFirebaseJWT, isAnonymousPortalData } from "../portal-api";
 import { IPortalData, IPortalDataUnion } from "../portal-types";
 
-export const getAttachmentsManagerOptions = async (portalData: IPortalDataUnion) => {
+export const getAttachmentsManagerOptions = async (portalData: IPortalDataUnion): Promise<IAttachmentsManagerInitOptions> => {
+
   const { basePortalUrl, rawPortalJWT } = portalData as IPortalData;
   let firebaseJwt: string | undefined;
   if (basePortalUrl && rawPortalJWT) {
